[1.0, 1.1] fixes #1511
git-svn-id: file:///var/www/production/svnmirror/svn_doctrine/branches/1.1/lib@5046 2b2ae520-0f61-44a6-8c59-7e116ffd6265

+++ Doctrine/Record.php
@@ -1541,7 +1541,9 @@ abstract class Doctrine_Record extends Doctrine_Record_Abstract implements Count
                     $this->refreshRelated($key);
                 }
 
-                $this->$key->fromArray($value, $deep);
+                if (is_array($value)) {
+                    $this->$key->fromArray($value, $deep);
+                }
             } else if ($this->getTable()->hasField($key)) {
                 $this->set($key, $value);
             } else {
@@ -1582,15 +1584,17 @@ abstract class Doctrine_Record extends Doctrine_Record_Abstract implements Count
                     $this->refreshRelated($key);
                 }
 
-                if (isset($value['_identifiers']) && is_array($value['_identifiers'])) {
-                    $this->unlink($key, array(), false);
-                    foreach ($value['_identifiers'] as $id => $exists) {
-                        if ($exists) {
-                            $this->link($key, $id, false);
+                if (is_array($value)) {
+                    if (isset($value['_identifiers']) && is_array($value['_identifiers'])) {
+                        $this->unlink($key, array(), false);
+                        foreach ($value['_identifiers'] as $id => $exists) {
+                            if ($exists) {
+                                $this->link($key, $id, false);
+                            }
                         }
+                    } else {
+                        $this->$key->synchronizeWithArray($value);
                     }
-                } else {
-                    $this->$key->synchronizeWithArray($value);
                 }
             } else if ($this->getTable()->hasField($key)) {
                 $this->set($key, $value);
